Remove fault context
